#42 create StyleParseTreeTransformer
