+++ api/parser/src/main/scala/de/htwg/zeta/parser/style/StyleParserImpl.scala
@@ -1,13 +1,7 @@
 package de.htwg.zeta.parser.style
 
-import javafx.scene.paint.Color
-
-import de.htwg.zeta.server.generator.model.style.color.{ColorOrGradient, ColorWithTransparency, Color => OldColor}
-import de.htwg.zeta.server.generator.model.style.gradient.GradientAlignment
-import de.htwg.zeta.server.generator.model.style.{Style, LineStyle => OldLineStyle}
-
-import scala.annotation.tailrec
-import scala.reflect.ClassTag
+import de.htwg.zeta.server.generator.model.style.color.{Color => OldColor}
+import de.htwg.zeta.server.generator.model.style.{LineStyle => OldLineStyle}
 
 
 class StyleParserImpl extends StyleParser {
@@ -17,66 +11,7 @@ class StyleParserImpl extends StyleParser {
   private val eq = literal("=")
   private val comma = literal(",")
 
-  override def styles: Parser[List[StyleParseTree]] = {
-    rep1(style).flatMap { styles =>
-      Parser { in =>
-        checkStyleRuleViolations(styles) match {
-          case Nil => Success(styles, in)
-          case styleRuleViolations => failureStyleRuleViolations(styleRuleViolations, in)
-        }
-      }
-    }
-  }
-
-  private def checkStyleRuleViolations(styles: List[StyleParseTree]): List[String] = {
-    val checks = List[List[StyleParseTree] => List[String]](
-      findStyleDuplicates,
-      findUndefinedParents,
-      findGraphCycles
-    )
-    checks.flatMap(_.apply(styles))
-  }
-
-  private def findStyleDuplicates(styles: List[StyleParseTree]): List[String] = {
-    val styleNames = styles.map(_.name)
-    val duplicates = styleNames.diff(styleNames.distinct)
-    duplicates.map(styleName => s"Style '$styleName' is defined multiple times (which is forbidden)")
-  }
-
-  private def findUndefinedParents(styles: List[StyleParseTree]): List[String] = {
-    val definedStyles = styles.map(_.name)
-    styles.flatMap(style => {
-      style.parentStyles.filter(parentStyle => !definedStyles.contains(parentStyle))
-        .map(undefinedStyle => s"Style '${style.name}' extends unknown style '$undefinedStyle'")
-    })
-  }
-
-  private def findGraphCycles(styles: List[StyleParseTree]): List[String] = {
-    for {
-      style <- styles if containsGraphCycles(style, styles)
-    } yield s"Found cyclic inheritance in style ${style.name}"
-  }
-
-  private def containsGraphCycles(style: StyleParseTree, allStyles: List[StyleParseTree]): Boolean = {
-    val visited = collection.mutable.Set[String]()
-
-    @tailrec
-    def check(parentNames: List[String]): Boolean = parentNames match {
-      case Nil => false
-      case _ if parentNames.contains(style.name) => true
-      case _ =>
-        val unvisited = parentNames.filterNot(visited.contains)
-        visited ++= unvisited
-        val parentStyles = unvisited.flatMap(parentName => allStyles.find(_.name == parentName))
-        val grandparents = parentStyles.flatMap(_.parentStyles)
-        check(grandparents)
-    }
-
-    check(style.parentStyles)
-  }
-
-  private def failureStyleRuleViolations(inheritanceRuleViolations: List[String], in: Input) = Failure(
-    s"The specified styles violate the inheritance rules: '${inheritanceRuleViolations.mkString(", ")}'", in)
+  override def styles: Parser[List[StyleParseTree]] = rep1(style)
 
   private def style: Parser[StyleParseTree] = {
     name ~ opt(parentStyles) ~ leftBraces ~ description ~ attributes ~ rightBraces ^^ { parseSeq =>
@@ -93,28 +28,7 @@ class StyleParserImpl extends StyleParser {
   private def attributes: Parser[List[StyleAttribute]] = {
     rep(lineColor | lineStyle | lineWidth | transparency | backgroundColor | fontColor | fontName | fontSize
       | fontBold | fontItalic | gradientOrientation | gradientAreaColor | gradientAreaOffset)
-      .flatMap { attributes =>
-        Parser { in =>
-          findAttributeDuplicates(attributes) match {
-            case Nil => Success(attributes, in)
-            case duplicateAttributes => failureDuplicateAttributes(duplicateAttributes, in)
-          }
-        }
-      }
   }
-
-  def findAttributeDuplicates(attributeList: List[StyleAttribute]): List[String] = {
-    val duplicates = attributeList.groupBy(_.attributeName).collect {
-      case (attributeName, attributes) if attributes.size > 1 => attributeName
-    }
-    duplicates.toList.sorted
-  }
-
-  private def failureDuplicateAttributes(duplicates: List[String], in: Input) = Failure(
-    s"""
-       |The specified style contains multiple occurrences of the following attributes (which is not allowed):"
-       |'${duplicates.mkString(", ")}'
-    """.stripMargin, in)
 
   private def name = literal("style") ~> ident
 
@@ -148,61 +62,3 @@ class StyleParserImpl extends StyleParser {
 
   private def parentStyles = literal("extends") ~> ident ~ rep(comma ~> ident) ^^ (parents => parents._1 :: parents._2)
 }
-
-object StyleParserImpl {
-
-  private trait ColorToRBGColor {
-    val color: Color
-
-    val getRGBValue: String = {
-      val r = color.getRed * 255.0.round.toInt
-      val g = color.getGreen * 255.0.round.toInt
-      val b = color.getBlue * 255.0.round.toInt
-
-      s"$r$g$b"
-    }
-  }
-
-  private case class ColorOrGradientImpl(color: Color) extends ColorOrGradient with ColorToRBGColor
-
-  private case class ColorWithTransparencyImpl(color: Color) extends ColorWithTransparency with ColorToRBGColor
-
-  private case class ColorImpl(color: Color) extends OldColor with ColorToRBGColor
-
-  def convert(styleParseTree: StyleParseTree): Style = {
-
-    class CollectAttributeWrapper[T](val t: Option[T]) {
-      def map[R](func: T => R): Option[R] = t.map(func)
-    }
-
-    def collectAttribute[T: ClassTag]: CollectAttributeWrapper[T] = {
-      val attribute = styleParseTree.attributes.collectFirst {
-        case t: T => t
-      }
-      new CollectAttributeWrapper(attribute)
-    }
-
-    new Style(
-      name = styleParseTree.name,
-      description = Some(styleParseTree.description),
-      transparency = collectAttribute[Transparency].map(_.transparency),
-      background_color = collectAttribute[BackgroundColor].map(bg => ColorOrGradientImpl(bg.color)),
-      line_color = collectAttribute[LineColor].map(lc => ColorWithTransparencyImpl(lc.color)),
-      line_style = collectAttribute[LineStyle].map(_.style).flatMap(OldLineStyle.getIfValid),
-      line_width = collectAttribute[LineWidth].map(_.width),
-      font_color = collectAttribute[FontColor].map(fc => ColorImpl(fc.color)),
-      font_name = collectAttribute[FontName].map(_.name),
-      font_size = collectAttribute[FontSize].map(_.size),
-      font_bold = collectAttribute[FontBold].map(_.bold),
-      font_italic = collectAttribute[FontItalic].map(_.italic),
-      gradient_orientation = collectAttribute[GradientOrientation].map(_.orientation).flatMap(GradientAlignment.ifValid),
-      selected_highlighting = None,
-      multiselected_highlighting = None,
-      allowed_highlighting = None,
-      unallowed_highlighting = None,
-      parents = List()
-    )
-  }
-
-}
-
