+++ api/parser/src/test/scala/de/htwg/zeta/parser/style/StyleParserTransformerTest.scala
@@ -5,7 +5,7 @@ import de.htwg.zeta.server.generator.model.style.gradient.VERTICAL
 import org.scalatest.FlatSpec
 import org.scalatest.Matchers
 
-class StyleParserConverterTest extends FlatSpec with Matchers {
+class StyleParserTransformerTest extends FlatSpec with Matchers {
 
   val parserToTest: StyleParserImpl = new StyleParserImpl
 
@@ -33,8 +33,9 @@ class StyleParserConverterTest extends FlatSpec with Matchers {
 
     val styleParser = parserToTest.parseStyles(styleToTestSuccess)
     styleParser.successful shouldBe true
-    val style: StyleParseTree = styleParser.get.head
-    val deprecatedStyle: Style = StyleParserImpl.convert(style)
+    val styleTrees = styleParser.get
+    val styles = StyleParseTreeTransformer.transform(styleTrees).getOrElse(List())
+    val deprecatedStyle: Style = styles.head
 
     deprecatedStyle.description shouldBe Some("\"Style for a connection between an interface and its implementing class\"")
     deprecatedStyle.line_width shouldBe Some(1)
